change schemas & add cloudinary

+++ src/App.Api/Controllers/ProductsController.cs
@@ -1,7 +1,10 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using App.Application.Contracts;
+using App.Api.Contracts;
 using App.Core.Abstractions;
 using App.Core.Models;
+using CloudinaryDotNet;
+using CloudinaryDotNet.Actions;
+using App.Application.Services;
 
 namespace App.Api.Controllers
 {
@@ -10,11 +13,16 @@ namespace App.Api.Controllers
 
     public class ProductsController : Controller
     {
+        private readonly IConfiguration _configuration;
         private readonly IProductsService _productsService;
+        private readonly Cloudinary _cloudinary;
 
-        public ProductsController(IProductsService productsService)
+        public ProductsController(IConfiguration configuration, IProductsService productsService)
         {
+            _configuration = configuration;
             _productsService = productsService;
+            _cloudinary = new Cloudinary(_configuration.GetConnectionString("CloudinaryUrl") ?? "");
+            _cloudinary.Api.Secure = true;
         }
 
         [HttpGet]
@@ -27,13 +35,18 @@ namespace App.Api.Controllers
                 p.Id,
                 p.Title,
                 p.Description,
-                p.OnSale,
+                p.Discount,
                 p.Rating,
                 p.Price,
                 DateTime.Now.AddDays(random.Next(7, 61)),
                 p.Brand,
                 p.IsFavorite,
-                p.Categories
+                p.Categories,
+                p.CreatedDate,
+                p.ChangedDate,
+                p.ImgUrl,
+                p.ImgPreviewUrl,
+                p.ImgId
                 ));
 
             return Ok(productsResponse);
@@ -55,30 +68,67 @@ namespace App.Api.Controllers
                 product.Id,
                 product.Title,
                 product.Description,
-                product.OnSale,
+                product.Discount,
                 product.Rating,
                 product.Price,
                 DateTime.Now.AddDays(random.Next(7, 61)),
                 product.Brand,
                 product.IsFavorite,
-                product.Categories
+                product.Categories,
+                product.CreatedDate,
+                product.ChangedDate,
+                product.ImgUrl,
+                product.ImgPreviewUrl,
+                product.ImgId
                 );
         }
 
         [HttpPost]
-        public async Task<ActionResult<Guid>> CreateProduct([FromBody] ProductsRequest request)
+        public async Task<ActionResult<Guid>> CreateProduct([FromForm] ProductsRequest request)
         {
+            string imageUrl = "";
+            string imgPreviewFile = "";
+            string? imgId = null;
+
+            if (request.ImageFile != null)
+            {
+                var uploadParams = new ImageUploadParams()
+                {
+                    File = new FileDescription(request.ImageFile.FileName, request.ImageFile.OpenReadStream()),
+                    UseFilename = true,
+                    UniqueFilename = true,
+                    Folder = "rida/products/"
+                };
+
+                var uploadResult = _cloudinary.Upload(uploadParams);
+
+                imgId = uploadResult.PublicId;
+                imageUrl = FileTransformationUrlService.GetTransformationUrl(
+                    uploadResult.SecureUrl.ToString(),
+                    ""
+                );
+                imgPreviewFile = FileTransformationUrlService.GetTransformationUrl(
+                    uploadResult.SecureUrl.ToString(),
+                    "c_thumb,h_200,w_200"
+                );
+            }
+
             var (product, error) = Product.Create(
                 Guid.NewGuid(),
                 request.Title,
                 request.Description,
-                request.OnSale,
+                request.Discount,
                 request.Rating,
                 request.Price,
                 request.Brand,
                 request.IsFavorite,
                 request.Categories,
-                request.IsHidden
+                request.IsHidden,
+                DateTime.UtcNow,
+                DateTime.UtcNow,
+                imageUrl,
+                imgPreviewFile,
+                imgId
                 );
 
             if (!string.IsNullOrEmpty(error))
@@ -92,19 +142,63 @@ namespace App.Api.Controllers
         }
 
         [HttpPut("{id:guid}")]
-        public async Task<ActionResult<Guid>> UpdateProduct(Guid id,[FromBody] ProductsRequest request)
+        public async Task<ActionResult<Guid>> UpdateProduct(Guid id,[FromForm] ProductsRequest request)
         {
+            var foundProduct = await _productsService.GetProductById(id);
+
+            if (foundProduct == null)
+            {
+                return NotFound($"Product with ID {id} not found.");
+            }
+
+            string imageUrl = foundProduct.ImgUrl;
+            string imgPreviewFile = foundProduct.ImgPreviewUrl;
+            string? imgId = foundProduct.ImgId;
+
+            if (request.ImageFile != null)
+            {
+                var uploadParams = new ImageUploadParams()
+                {
+                    File = new FileDescription(request.ImageFile.FileName, request.ImageFile.OpenReadStream()),
+                    UseFilename = true,
+                    UniqueFilename = true,
+                    Folder = "rida/products/"
+                };
+
+                var uploadResult = _cloudinary.Upload(uploadParams);
+
+                imgId = uploadResult.PublicId;
+                imageUrl = FileTransformationUrlService.GetTransformationUrl(
+                    uploadResult.SecureUrl.ToString(),
+                    ""
+                );
+                imgPreviewFile = FileTransformationUrlService.GetTransformationUrl(
+                    uploadResult.SecureUrl.ToString(),
+                    "c_thumb,h_200,w_200"
+                );
+
+                if (foundProduct.ImgId != null)
+                {
+                    var deletionParams = new DeletionParams(foundProduct.ImgId);
+                    var deletionResult = _cloudinary.Destroy(deletionParams);
+                }
+            }
+
             var productId = await _productsService.UpdateProduct(
                 id,
                 request.Title,
                 request.Description,
-                request.OnSale,
+                request.Discount,
                 request.Rating,
                 request.Price,
                 request.Brand,
                 request.IsFavorite,
                 request.Categories,
-                request.IsHidden
+                request.IsHidden,
+                DateTime.UtcNow,
+                imageUrl,
+                imgPreviewFile,
+                imgId
                 );
 
             return Ok(productId);
@@ -113,6 +207,20 @@ namespace App.Api.Controllers
         [HttpDelete("{id:guid}")]
         public async Task<ActionResult<Guid>> DeleteProduct(Guid id)
         {
+            var foundProduct = await _productsService.GetProductById(id);
+
+            if (foundProduct == null)
+            {
+                return NotFound($"Product with ID {id} not found.");
+            }
+
+            if (!string.IsNullOrEmpty(foundProduct.ImgId))
+            {
+                var deletionParams = new DeletionParams(foundProduct.ImgId);
+
+                var deletionResult = _cloudinary.Destroy(deletionParams);
+            }
+
             var productId = await _productsService.DeleteProduct( id );
 
             return Ok(productId);
